some fixes

+++ app/api/v1/users.py
@@ -1,8 +1,6 @@
 from typing import List, Optional, Union
 from uuid import UUID
 
-from async_fastapi_jwt_auth import AuthJWT
-from async_fastapi_jwt_auth.exceptions import MissingTokenError
 from fastapi import APIRouter, Depends, HTTPException, status
 from pydantic import conint
 from schemas.auth import RefreshToken, TwoTokens
@@ -10,20 +8,14 @@ from schemas.base import HTTPExceptionResponse, HTTPValidationError
 from schemas.role import RoleBaseUUID
 from schemas.session import SessionResponse
 from schemas.user import UserPatch, UserResponse
+from services.auth import AuthService, get_auth_service
 from services.session import SessionService, get_session_service
 from services.user import UserService, get_user_service
 
 router = APIRouter()
 
-async def require_jwt(auth_jwt: AuthJWT):
-    try:
-        await auth_jwt.jwt_required()
-    except MissingTokenError:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Missing token")
-
 @router.delete(
     "/sessions/{session_id}",
-    response_model=None,
     summary="Delete user session",
     responses={
         "401": {"model": HTTPExceptionResponse},
@@ -35,11 +27,13 @@ async def require_jwt(auth_jwt: AuthJWT):
 async def delete_user_session(
     session_id: UUID,
     session_service: SessionService = Depends(get_session_service),
-    auth_jwt: AuthJWT = Depends(require_jwt)
+    auth_service: AuthService = Depends(get_auth_service)
 ) -> Optional[Union[HTTPExceptionResponse, HTTPValidationError]]:
     """
     Delete user session by session ID.
     """
+    await auth_service.check_access()  # Ensure the user is authenticated
+
     session = await session_service.get_session(session_id)
     if not session:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Session not found")
@@ -64,11 +58,13 @@ async def get_user_sessions(
     page_size: PageSizeType = 50,
     page_number: PageSizeType = 1,
     session_service: SessionService = Depends(get_session_service),
-    auth_jwt: AuthJWT = Depends(require_jwt)
+    auth_service: AuthService = Depends(get_auth_service)
 ) -> Union[List[SessionResponse], HTTPExceptionResponse]:
     """
     Retrieve user's session history with optional pagination and activity filter.
     """
+    await auth_service.check_access()  # Ensure the user is authenticated
+
     # Simulate session retrieval from cache or database using the service
     sessions = await session_service.get_all_sessions()  # Implement in service
     if not sessions:
@@ -96,13 +92,14 @@ async def add_role_to_user(
     user_id: UUID, 
     body: RoleBaseUUID, 
     user_service: UserService = Depends(get_user_service),
-    auth_jwt: AuthJWT = Depends(require_jwt)
+    auth_service: AuthService = Depends(get_auth_service)
 ) -> Optional[Union[HTTPExceptionResponse, HTTPValidationError]]:
     """
     Add a role to a user.
     """
+    await auth_service.check_access()  # Ensure the user is authenticated
+
     try:
-        # maybe this should be in role service?
         await user_service.add_role_to_user(user_id, body) 
     except Exception as e:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(e))
@@ -125,13 +122,14 @@ async def take_away_role_from_user(
     user_id: UUID, 
     body: RoleBaseUUID, 
     user_service: UserService = Depends(get_user_service),
-    auth_jwt: AuthJWT = Depends(require_jwt)
+    auth_service: AuthService = Depends(get_auth_service)
 ) -> Optional[Union[HTTPExceptionResponse, HTTPValidationError]]:
     """
     Remove a role from a user.
     """
+    await auth_service.check_access()  # Ensure the user is authenticated
+
     try:
-        # maybe this should be in role service?
         await user_service.remove_role_from_user(user_id, body)
     except Exception as e:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e))
@@ -150,11 +148,13 @@ async def take_away_role_from_user(
 )
 async def get_user_info(
     user_service: UserService = Depends(get_user_service),
-    auth_jwt: AuthJWT = Depends(require_jwt),
+    auth_service: AuthService = Depends(get_auth_service),
 ) -> Union[UserResponse, HTTPExceptionResponse]:
     """
     Retrieve current user's information.
     """
+    await auth_service.check_access()  # Ensure the user is authenticated
+
     user = await user_service.get_current_user()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
@@ -174,11 +174,13 @@ async def get_user_info(
 async def patch_current_user(
     body: UserPatch,
     user_service: UserService = Depends(get_user_service),
-    auth_jwt: AuthJWT = Depends(require_jwt),
+    auth_service: AuthService = Depends(get_auth_service),
 ) -> Union[UserResponse, HTTPExceptionResponse, HTTPValidationError]:
     """
     Update the current user's profile.
     """
+    await auth_service.check_access()  # Ensure the user is authenticated
+
     try:
         updated_user = await user_service.update_user(body)
     except Exception as e:
